Merge pull request #41 from retrozinndev/feat/comp-config-injection

+++ install.sh
@@ -3,11 +3,11 @@
 trap "printf \"\nOk, quitting beacuse you entered an exit signal. (SIGINT).\n\"; exit 1" SIGINT
 trap "printf \"\nOh noo!! Some application just killed the script! (SIGTERM)\"; exit 2" SIGTERM
 
-XDG_DATA_HOME=`[[ -z "$XDG_DATA_HOME" ]] && echo -n "$HOME/.local/share" || echo -n "$XDG_DATA_HOME"`
-XDG_CACHE_HOME=`[[ -z "$XDG_CACHE_HOME" ]] && echo -n $HOME/.cache || echo -n $XDG_CACHE_HOME`
-XDG_CONFIG_HOME=`[[ -z "$XDG_CONFIG_HOME" ]] && echo -n "$HOME/.config" || echo -n "$XDG_CONFIG_HOME"`
-BIN_HOME=`[[ -z "$BIN_HOME" ]] && echo -n "$HOME/.local/bin" || echo -n "$BIN_HOME"`
-APPS_HOME=`[[ -z "$APPS_HOME" ]] && echo -n "$XDG_DATA_HOME/applications" || echo -n "$APPS_HOME"`
+XDG_DATA_HOME=${XDG_DATA_HOME:-"$HOME/.local/share"}
+XDG_CACHE_HOME=${XDG_CACHE_HOME:-"$HOME/.cache"}
+XDG_CONFIG_HOME=${XDG_CONFIG_HOME:-"$HOME/.config"}
+BIN_HOME=${BIN_HOME:-"$HOME/.local/bin"}
+APPS_HOME=${APPS_HOME:-"$XDG_DATA_HOME/applications"}
 
 skip_prompts=`[[ "$1" == -y ]] && echo -n true`
 is_standalone=`(git remote -v > /dev/null 2>&1) || echo -n true`
@@ -45,16 +45,7 @@ sleep .5
 echo "Welcome to the colorshell installation script!"
 
 # Warn user of possible issues
-Send_log warn "!! By running this script, you assume total responsability for any issues that may occur with your filesystem"
-
-[[ -z $skip_prompts ]] && \
-    Send_log warn "Your current Hyprland and kitty configuration will be overwritten, accept the backup prompt if you still want them"
-
-[[ -z $skip_prompts ]] && \
-    Ask "Do you want to backup what is going to be modified/overwritten?"
-
-[[ $answer == y ]] || [[ $skip_prompts ]] && \
-    Backup_config $repo_directory
+Send_log warn "!! By running this, you're assuming total responsability for any issues that may occur with your filesystem"
 
 [[ -z "$skip_prompts" ]] && \
     Ask "Do you want to start the shell installation?"
@@ -65,10 +56,10 @@ if [[ "$answer" == y ]] || [[ "$skip_prompts" ]]; then
         rm -rf $repo_directory 2> /dev/null
         Send_log "Cloning repository in \`$repo_directory\`..."
         if [[ -d $repo_directory/.git ]]; then
-            Send_log "repo is already cloned! let's just fetch the latest changes..."
+            Send_log "Repo is already cloned! let's just fetch the latest changes..."
             git -C "$repo_directory" stash # if there are changes, let's just stash them
             git -C "$repo_directory" checkout ryo
-            git -C "$repo_directory" fetch && git -C "$repo_directory" pull --rebase
+            git -C "$repo_directory" fetch && git -C "$repo_directory" pull --rebase # rebase just in case
         else
             git clone https://github.com/retrozinndev/colorshell.git "$repo_directory"
         fi
@@ -77,7 +68,7 @@ if [[ "$answer" == y ]] || [[ "$skip_prompts" ]]; then
     Ask "Nice! Do you want to use the stable version instead of the unstable(latest commit)?"
 
     if [[ -z "$skip_prompts" ]] && [[ "$answer" == y ]]; then
-        Send_log "fetching latest release from colorshell repository"
+        Send_log "Fetching latest release from colorshell repository"
         latest_tag=`curl -s "$repo_api_url/releases" | jq -r '. | select(.[].prerelease == false) | .[0].tag_name'`
         
         Send_log "Done fetching"
@@ -85,23 +76,9 @@ if [[ "$answer" == y ]] || [[ "$skip_prompts" ]]; then
         git -C "$repo_directory" checkout $latest_tag > /dev/null 2>&1
     fi
 
-    Send_log "Starting installation..."
-
-    Send_log "Installing default configurations"
-    for dir in $(ls -A -w1 "$repo_directory/config"); do
-        dest=$XDG_CONFIG_HOME/$dir
-
-        Send_log "Installing $dir in $dest"
-        mkdir -p `dirname "$dest"` # create parents
-
-        [[ -d "$dest" ]] || [[ -f "$dest" ]] && \
-            rm -rf $dest
-
-        cp -rf $repo_directory/config/$dir "$dest" # copy
-    done
-
+    Send_log "Starting build process..."
     Send_log "Updating dependencies"
-    pnpm -C "$repo_directory" i && pnpm -C "$repo_directory" update
+    pnpm -C "$repo_directory" i && pnpm -C "$repo_directory" update > /dev/null
 
     Send_log "Building colorshell"
     pnpm -C "$repo_directory" build:release
@@ -119,10 +96,28 @@ if [[ "$answer" == y ]] || [[ "$skip_prompts" ]]; then
     mkdir -p $APPS_HOME
     cp -f $repo_directory/build/release/colorshell.desktop $APPS_HOME
 
+    # Check if user has a Hyprland config file
+    if [[ ! -f "$XDG_CONFIG_HOME/hypr/hyprland.conf" ]] && [[ -f "/usr/share/hypr/hyprland.conf" ]]; then
+        Send_log "Looks like Hyprland wasn't launched yet! Copying default config file..."
+        mkdir -p "$XDG_CONFIG_HOME/hypr"
+        cp -f "/usr/share/hypr/hyprland.conf" "$XDG_CONFIG_HOME/hypr/hyprland.conf"
+        Send_log "Adding exec for colorshell in Hyprland config..."
+        echo -ne "\nexec-once = ~/.local/bin/colorshell" >> "$XDG_CONFIG_HOME/hypr/hyprland.conf"
+    else
+        Ask "Do you want to autostart colorshell with Hyprland?"
+        if [ "$answer" == "y" ]; then
+            Send_log "Adding exec-once for colorshell to Hyprland..."
+            echo -ne "\nexec-once = ~/.local/bin/colorshell" >> "$XDG_CONFIG_HOME/hypr/hyprland.conf"
+        fi
+    fi
 
-    Send_log "Adding default wallpaper in ~/wallpapers"
-    mkdir -p $HOME/wallpapers
-    cp -f $repo_directory/resources/wallpaper_default.jpg "$HOME/wallpapers/Default Hypr-chan.jpg"
+    # Check for wallpaper configuration
+    if [[ ! -f "$XDG_CONFIG_HOME/hypr/hyprpaper.conf" ]]; then
+        Send_log "No hyprpaper config found, using colorshell's default..."
+        mkdir -p $HOME/wallpapers
+        cp -f $repo_directory/resources/wallpaper_default.jpg "$HOME/wallpapers/Default Hypr-chan.jpg"
+        cp -f $repo_directory/resources/config/hyprpaper.conf "$XDG_CONFIG_HOME/hypr/hyprpaper.conf"
+    fi
 
     if [[ -z "$skip_prompts" ]]; then
         echo "Colorshell is installed! :D"
